moved rounding tests
